separate vector processing from raster processing
